requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
PyMySQL
python-dotenv